requirements.txt
opencv-python-headless
numpy
tensorflow
google-generativeai
Pillow
